+++ love/src/Config.js
@@ -653,6 +653,28 @@ export const ccCameraFilterChangerDetailedStateToStyle = {
 export const alignedStateToStyle = {
   ALIGNED: 'ok',
   'NOT ALIGNED': 'warning',
+}
+
+//DM Flow
+export const dmFlowStatusMap = {
+  1: 'INTEGRATING',
+  2: 'FORDWARDED',
+  3: 'STOPPED',
+  4: 'DROPPED',
+  5: 'COMPLETE',
+  6: 'COPYING',
+  7: 'PROCESSED',
+  0: 'UNKNOWN',
+};
+
+export const stateToStyleDMFlow = {
+  INTEGRATING: 'warning',
+  FORDWARDED: 'warning',
+  STOPPED: 'ok',
+  DROPPED: 'alert',
+  COMPLETE: 'running',
+  COPYING: 'warning',
+  PROCESSED: 'running',
   UNKNOWN: 'invalid',
 };
 
@@ -2653,6 +2675,50 @@ export const signalBypassIndexes = {
   notParkedAtTheTelescopeLevel: [25, 15],
   interlock: [26, 1],
   manLiftNotParked: [26, 2],
+};
+
+/***************************************************************************/
+/***************************** Dynalenel System ***************************/
+/*************************************************************************/
+
+export const dynaleneDetailedStateMap = {
+  0: 'Initialized',
+  1: 'ShuttingDown',
+  2: 'PoweringOn',
+  3: 'PoweredOn',
+  4: 'PoweringOff',
+  5: 'PoweredOff',
+  6: 'Warning',
+  7: 'Alarm',
+  8: 'ShutOff',
+  9: 'Unknown',
+};
+
+export const dynaleneDetailedStateToStyle = {
+  Initialized: 'ok',
+  ShuttingDown: 'warning',
+  PoweringOn: 'warning',
+  PoweredOn: 'warning',
+  PoweringOff: 'warning',
+  PoweredOff: 'undefined',
+  Warning: 'warning',
+  Alarm: 'alert',
+  ShutOff: 'undefined',
+  Unknown: 'undefined',
+};
+
+export const dynaleneTankLevelDetailedStateMap = {
+  0: 'Ok',
+  1: 'Warning',
+  2: 'Alarm',
+  9: 'Unknown',
+};
+
+export const dynaleneTankLevelToStyle = {
+  Ok: 'ok',
+  Warning: 'warning',
+  Alarm: 'alert',
+  Unknown: 'invalid',
 };
 
 /*****************************************************************************/