+++ love/src/components/GeneralPurpose/Plot/PolarPlot/PolarPlot.container.jsx
@@ -77,9 +77,9 @@ export const schema = {
       default: {
         WindSpeed: {
           category: 'telemetry',
-          csc: 'WeatherStation',
-          salindex: '1',
-          topic: 'windSpeed',
+          csc: 'ESS',
+          salindex: 301,
+          topic: 'airFlow',
           item: 'avg2M',
           type: 'line',
           accessor: '(x) => x',
@@ -89,9 +89,9 @@ export const schema = {
         },
         WindDirection: {
           category: 'telemetry',
-          csc: 'WeatherStation',
-          salindex: '1',
-          topic: 'windDirection',
+          csc: 'ESS',
+          salindex: 301,
+          topic: 'airFlow',
           item: 'avg2M',
           type: 'line',
           accessor: '(x) => x',
@@ -126,8 +126,8 @@ export const schema = {
       type: 'function',
       description: 'Color interpolation function. It takes a value, its range (min and max) and the group index.',
       default: `(value, minValue, maxValue, group) => { 
-  const proportion = (value - minValue) / (maxValue - minValue); 
-  return [255, 255 * (1 - proportion), 255 * (1 - proportion)]; 
+  const proportion = maxValue !== minValue ? (value - minValue) / (maxValue - minValue) : 0;
+  return [255 * (1 - proportion), 255, 255 * (1 - proportion)];
 }`,
       isPrivate: false,
     },