+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Version History
 ===============
 
+v5.27.0
+-------
+
+* DM Flow implementation `<https://github.com/lsst-ts/LOVE-frontend/pull/554>`_
+* Point LOVE weather station to read corresponding CSC `<https://github.com/lsst-ts/LOVE-frontend/pull/553>`_
+* Add Dynalene Component `<https://github.com/lsst-ts/LOVE-frontend/pull/552>`_
+* Add new params: category and time_lost_type to narrative log `<https://github.com/lsst-ts/LOVE-frontend/pull/551>`_
+
 v5.26.1
 -------